requirements.txt
streamlit
google.generativeai
plotly
kaleido
openpyxl
python-dotenv